splitting trainer modules into two parts
